chore: 스토리북 파일 GNB와 Fallback 합쳐서 components로 분류
