--- src/components/GNB/Header.stories.tsx
@@ -1,18 +1,0 @@
-import Header from "./Header";
-import { MemoryRouter } from "react-router-dom";
-import type { Meta } from "@storybook/react";
-
-const meta: Meta = {
-  title: "gnb/header",
-  component: Header,
-};
-
-export default meta;
-
-export const Default = {
-  render: () => (
-    <MemoryRouter>
-      <Header />
-    </MemoryRouter>
-  ),
-};
