--- src/components/fallback/Blank/index.stories.tsx
@@ -1,13 +1,0 @@
-import Blank from ".";
-import type { Meta } from "@storybook/react";
-
-const meta: Meta = {
-  title: "fallback/blank",
-  component: Blank,
-};
-
-export default meta;
-
-export const Default = {
-  args: {},
-};
