--- src/components/fallback/ErrorFallback/index.stories.tsx
@@ -1,13 +1,0 @@
-import ErrorFallback from ".";
-import type { Meta } from "@storybook/react";
-
-const meta: Meta = {
-  title: "fallback/errorFallback",
-  component: ErrorFallback,
-};
-
-export default meta;
-
-export const Default = {
-  args: {},
-};
