--- src/components/fallback/LoadingFallback/index.stories.tsx
@@ -1,13 +1,0 @@
-import LoadingFallback from ".";
-import type { Meta } from "@storybook/react";
-
-const meta: Meta = {
-  title: "fallback/loadingFallback",
-  component: LoadingFallback,
-};
-
-export default meta;
-
-export const Default = {
-  args: {},
-};
